manage like and dislike

+++ controllers/sauces.js
@@ -1,3 +1,4 @@
+// import models BDD from file models
 const Sauce = require("../models/sauces");
 // import file systeme to delete file in like our file images
 const fs = require("fs"); // systeme interne de node
@@ -63,8 +64,19 @@ exports.getAllSauce = (req, res, next) => {
 };
 
 exports.deleteSauce = (req, res, next) => {
-  // before delete object from our database we need to know his URL
+  // verif if the user whowant to do a request is the same user that we athentify with our database
   Sauce.findOne({ _id: req.params.id }).then((sauce) => {
+    if (!sauce) {
+      return res.status(404).json({
+        error: new Error("Sauce non trouvé !"),
+      });
+    }
+    if (sauce.userId !== req.auth.userId) {
+      return res.status(401).json({
+        error: new Error("requête non autorisé"),
+      });
+    }
+    // before delete object from our database we need to know his URL
     const filename = sauce.imageUrl.split("/images/")[1]; // url/images/namefile.jpg ==> we need the name of file to delete it
     fs.unlink(`images/${filename}`, () => {
       Sauce.deleteOne({ _id: req.params.id })
@@ -82,31 +94,45 @@ exports.deleteSauce = (req, res, next) => {
     });
   });
 
-  // verif if the user whowant to do a request is the same user that we athentify with our database
-  Sauce.findOne({ _id: req.params.id }).then((sauce) => {
-    if (!sauce) {
-      return res.status(404).json({
-        error: new Error("Sauce non trouvé !"),
-      });
-    }
-    if (sauce.userId !== req.auth.userId) {
-      return res.status(401).json({
-        error: new Error("requête non autorisé"),
-      });
-    }
-    // after matching userid with our data base we can produce to response to request delete
-    // to delete it we move this one to our callaback fs.unlink
-    // Sauce.deleteOne({ _id: req.params.id })
-    //   .then(() => {
-    //     //delete sauce
-    //     res.status(200).json({
-    //       message: "sauce supprimé!",
-    //     });
-    //   })
-    //   .catch((error) => {
-    //     res.status(400).json({
-    //       error: error,
-    //     });
-    //   });
-  });
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+  // Sauce.findOne({ _id: req.params.id }).then((sauce) => {
+  //   if (!sauce) {
+  //     return res.status(404).json({
+  //       error: new Error("Sauce non trouvé !"),
+  //     });
+  //   }
+  //   if (sauce.userId !== req.auth.userId) {
+  //     return res.status(401).json({
+  //       error: new Error("requête non autorisé"),
+  //     });
+  //   }
+  //   // after matching userid with our data base we can produce to response to request delete
+  //   // to delete it we move this one to our callaback fs.unlink
+  //   // Sauce.deleteOne({ _id: req.params.id })
+  //   //   .then(() => {
+  //   //     //delete sauce
+  //   //     res.status(200).json({
+  //   //       message: "sauce supprimé!",
+  //   //     });
+  //   //   })
+  //   //   .catch((error) => {
+  //   //     res.status(400).json({
+  //   //       error: error,
+  //   //     });
+  //   //   });
+  // });
 };
+
